--- a/lib/import/brca/providers/oxford/oxford_handler.rb
+++ b/lib/import/brca/providers/oxford/oxford_handler.rb
@@ -11,7 +11,8 @@ module Import
 
             return if @files_not_to_process.include? @file_name
 
-            if brca_file?
+            if @files_to_process.include?(@file_name) || brca_file?
+              @files_to_process << @file_name unless @files_to_process.include?(@file_name)
               prepare_genotypes(record)
             else
               @files_not_to_process << @file_name