## What?

Oxford's scripts were too slow and taking hours to finish up , making it hard to bring any changes as everytime a lot of developer's time was going in executing it. More details on planio - 36735

## Why?
For every record coming , check was being made if the file it belongs to is relevant to process under current handler or not.

## How?
I have checked on first incoming record , if the file should be processed or not and return early if its not to be processed rather checking everytime for each record.

## Testing ?
Locally tested via timing the script and here are the changes in timings - 

Old - 
BRCA script finishes at - Total time taken - 0 hours, 55 minutes, and 51 seconds.
CRC script finishes at - Total time taken - 0 hours, 40 minutes, and 19 seconds.

New times -
CRC - Total time taken - 0 hours, 10 minutes, and 38 seconds.
BRCA - Total time taken - 0 hours, 11 minutes, and 19 seconds.

Also compared the counts before and after changes and they stay the same.